Refactor pagination
