+++ src/components/layout/Pagination.jsx
@@ -1,19 +1,19 @@
 import React, { useState, useEffect, useContext } from "react";
-import CryptoContext from "../context/CryptoContext";
+import CryptoContext from "../../context/CryptoContext";
 
 function Pagination() {
   const { dispatch, currentPage, currentPageAssets, assets } =
     useContext(CryptoContext);
   const pageNumbers = [];
-
+  const assetsPerPage = 8;
+  const totalAssets = assets.length;
   useEffect(() => {
     const indexOfLastAsset = currentPage * assetsPerPage;
     const indexOfFirstAsset = indexOfLastAsset - assetsPerPage;
-    const currentPageAssets = assets.slice(indexOfFirstAsset, indexOfLastAsset);
-    dispatch({ type: "SET_CURRENT_PAGE_ASSETS", payload: currentPageAssets });
+    const currentPageItems = assets.slice(indexOfFirstAsset, indexOfLastAsset);
+    console.log(currentPageItems);
+    dispatch({ type: "SET_CURRENT_PAGE_ASSETS", payload: currentPageItems });
   }, [dispatch, currentPage]);
-  const assetsPerPage = 8;
-  const totalAssets = assets.length;
   for (let i = 1; i <= Math.ceil(totalAssets / assetsPerPage); i++) {
     pageNumbers.push(i);
   }
