+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import CryptoList from "../components/crypto/CryptoList";
-import Search from "../components/Search";
+import Search from "../components/layout/Search";
 import { useContext, useEffect } from "react";
 import CryptoContext from "../context/CryptoContext";
 import { getMultipleAssets } from "../context/CryptoActions";
